backend/requirements.txt
awscli==1.16.9
awsebcli==3.14.4
blessed==1.15.0
boto3==1.9.4
botocore==1.11.9
cached-property==1.4.3
cement==2.8.2
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-cors-headers==2.3.0
django-filter==1.1.0
django-oauth-toolkit==1.2.0
django-storages==1.7.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
docker==3.5.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.1.0
pathspec==0.5.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
PyJWT==1.6.4
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
PyYAML==3.13
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.51.0
